Fixed errors when tests are skipped in story.t
git-svn-id: http://svn.krangcms.com/trunk/krang@1664 689d98c0-9023-0410-9c0b-f79783857903

+++ t/story.t
@@ -819,10 +819,10 @@ sub test_linked_media {
     }
 
     END {
-        $media->delete();
-        $media2->delete();
-        $contrib->delete();
-        $story->delete();
+        $media->delete() if $media;
+        $media2->delete() if $media2;
+        $contrib->delete() if $contrib;
+        $story->delete() if $story;
     }
 
 }
